requirements.txt
notion_client==2.3.0
openai==1.75.0
python-dotenv==1.1.0
Requests==2.32.3
rich==14.0.0
tqdm==4.67.1
typer==0.15.2
